+++ pkg/httpx/httpx.go
@@ -1,9 +1,14 @@
 package httpx
 
 const (
-	HeaderContentType   = "Content-Type"
-	HeaderContentLength = "Content-Length"
-	HeaderContentRange  = "Content-Range"
-	HeaderRange         = "Range"
-	HeaderAcceptRanges  = "Accept-Ranges"
+	HeaderContentType     = "Content-Type"
+	HeaderContentLength   = "Content-Length"
+	HeaderContentRange    = "Content-Range"
+	HeaderRange           = "Range"
+	HeaderAcceptRanges    = "Accept-Ranges"
+	HeaderUserAgent       = "User-Agent"
+	HeaderAccept          = "Accept"
+	HeaderAuthorization   = "Authorization"
+	HeaderWWWAuthenticate = "WWW-Authenticate"
+	HeaderXForwardedFor   = "X-Forwarded-For"
 )